Inheritance compiles in LDC
Had to completely ignore my template wrapper to get the function pointer
type because LDC tries to export all C++ symbols which means even hiding
a function stub in a D template cannot be instantiated in C++ correctly.
So now there's annoying symbol pollution in the vtable/mtable objects.
Sigh. ALL BECAUSE I CAN'T DIRECTLY ALIAS A REF RETURN TYPE FOR A
FUNCTION POINTER WHY IS THIS SO HARD