binderoo_host_dotnet added. Currently supports creating a Host (required to launch binderoo) and Imported Functions.
Imported functions are a pain - thanks to Generics not handling attributes on types (sigh) and the marshaller not accepting delegate types that are generic in any way (double sigh) it relies on generated code by a binderoo client DLL to properly format and call the functions. This is rubbish, but that's C# for ya. See binderoo_util -ic#A output to see how ImportedFunction is used properly.