Part two of a rethink of the mangling hack.
Why go to all the trouble of hiding the functions when they're just going to show up on Linux? So I had a better idea. Why don't we generate the correct mangles for them so that you can static link them?

In reality this is just picking up an idea Manu (@turkeyman) was trying to implement in Remedy's original binding code - we don't necessarily want to have a full rapid iteration system in a retail product. So static linking everything we want with correct mangled symbols means we can go ahead get rid of the rapid iteration.

This mangler will generate the correct C and C++ manglings from any given function. With my limited test cases, that is. Function pointers are out of the question as types for the time being. As are templated types (ie DSlice< _ty > ) so strings aren't handled correctly. They'll come later, just commiting this for the time being.

Also, this is only a Windows implementation. GCC/LLVM implementation to come whenever I switch back over to Linux.