.NET project updated with added files. And another attempt to remove x86/anyplatform from the project. Why does it keep being re-added. Jeez. Stop it Microsoft.
Also explicitly link to the binderoo_host DLL so that it gets copied over on build. Which involved needing to hand edit the project file, because trying to add the file through the IDE dialogue would result in it being copied over for all of existence. Also, I conditionally choose debug or release this way. Sigh.