Binding is no longer done internally by mixing in a template to the binding module; and as a result, the BindModules mixin now is a thing. Additionally, BindModules is the first of the mixins that takes the new BindOption enumeration as a part of its template parameter tuple.
This was the result of Laeeth saying at DConf that he'd prefer a pull-in method of binding modules (backed up by showing some of Atila's work) than the push-in method that was currently being used. I decided to give it a try on some code I'm writing for my startup, and worked through all the issues. The code is slightly cleaner now as well.